Rename config.h to format_config.h after spurious (?) compilation complaint about src/include/config.h being shadowed.
